Add translation function for poll subjects
type:feat
body: used Googletrans module to translate from language of the poll to English. In some cases, the stated original language was inaccurate (French for what was in fact, English, and had to be overridden). The process took a few hours over 10,700 or so translations. Some redundancy could be removed because poll titles are repeated for different questions.